progress is returned as float with 2 decimal places

+++ engine-spark/src/main/scala/com/intel/intelanalytics/engine/spark/CommandProgressUpdater.scala
@@ -4,5 +4,5 @@ package com.intel.intelanalytics.engine.spark
  * Execute when receiving progress update for command
  */
 trait CommandProgressUpdater {
-  def updateProgress(commandId: Long, progress: List[Int])
+  def updateProgress(commandId: Long, progress: List[Float])
 }
